Added Human User ID, Tested & added Session tokens

+++ models/Record.go
@@ -3,7 +3,7 @@ package models
 import "github.com/Simon-Martens/caveman/tools/types"
 
 type Record struct {
-	ID       int            `db:"pk,id"`
+	ID       int64          `db:"pk,id"`
 	Created  types.DateTime `db:"created"`
 	Modified types.DateTime `db:"modified"`
 }
